bot: fix derps - add method to check if the link contains any file or not - fix filename derps

+++ bot.py
@@ -83,7 +83,7 @@ def start_bot(bot, update):
 	#site = httplib.HTTPConnection(url)
 	#site.request("HEAD", '')
 	#if site.getresponse().status == 200:
-	if validators.url(url):
+	if validators.url(url) or download.is_downloadable(url):
 		sent_message = bot.send_message(chat_id=update.message.chat_id, text=Text.PROCESSING)
 		if user_cmd:
 			if(user_cmd == "video"):
@@ -114,27 +114,33 @@ def start_bot(bot, update):
 				else:
 					sent_message.edit_text(Text.NOT_SUPPORTED,parse_mode=telegram.ParseMode.HTML)
 			else:
-				filename = user_cmd
-				raw_file = download.download(url, filename)
-				if "ERROR" in raw_file:
-					sent_message.edit_text(Text.FAILED+raw_file, parse_mode=telegram.ParseMode.HTML)
+				if download.is_downloadable(url):
+					filename = user_cmd
+					raw_file = download.download(url, filename)
+					if "ERROR" in raw_file:
+						sent_message.edit_text(Text.FAILED+raw_file, parse_mode=telegram.ParseMode.HTML)
+					else:
+						bot.send_chat_action(chat_id=update.message.chat_id, action=telegram.ChatAction.TYPING)
+						sent_message.edit_text(Text.UPLOADING_GD)
+						dwnld_url = upload.upload(raw_file)
+						size = (os.path.getsize(raw_file))/1048576
+						short_link = upload.shorten(dwnld_url)
+						sent_message.edit_text(Text.DONE.format(raw_file, size, short_link),parse_mode=telegram.ParseMode.HTML)
+						os.remove(raw_file)
 				else:
-					bot.send_chat_action(chat_id=update.message.chat_id, action=telegram.ChatAction.TYPING)
-					sent_message.edit_text(Text.UPLOADING_GD)
-					dwnld_url = upload.upload(raw_file)
-					size = (os.path.getsize(raw_file))/1048576
-					short_link = upload.shorten(dwnld_url)
-					sent_message.edit_text(Text.DONE.format(filename, size, short_link),parse_mode=telegram.ParseMode.HTML)
-					os.remove(raw_file)
+					sent_message.edit_text(Text.ISNOT_DOWNLOADABLE,parse_mode=telegram.ParseMode.HTML)
 		else:
-			raw_file = download.download(url, None)
-			bot.send_chat_action(chat_id=update.message.chat_id, action=telegram.ChatAction.TYPING)
-			sent_message.edit_text(Text.UPLOADING_GD)
-			dwnld_url = upload.upload(raw_file)
-			size = (os.path.getsize(raw_file))/1048576
-			short_link = upload.shorten(dwnld_url)
-			sent_message.edit_text(Text.DONE.format(filename, size, short_link),parse_mode=telegram.ParseMode.HTML)
-	 		os.remove(raw_file)
+			if download.is_downloadable(url):
+				raw_file = download.download(url, None)
+				bot.send_chat_action(chat_id=update.message.chat_id, action=telegram.ChatAction.TYPING)
+				sent_message.edit_text(Text.UPLOADING_GD)
+				dwnld_url = upload.upload(raw_file)
+				size = (os.path.getsize(raw_file))/1048576
+				short_link = upload.shorten(dwnld_url)
+				sent_message.edit_text(Text.DONE.format(raw_file, size, short_link),parse_mode=telegram.ParseMode.HTML)
+		 		os.remove(raw_file)
+			else:
+				sent_message.edit_text(Text.ISNOT_DOWNLOADABLE,parse_mode=telegram.ParseMode.HTML)
 	elif("help" not in url and "start" not in url and "broadcast" not in url and "donate" not in url and "add_user" not in url and "revoke_user" not in url):
 		bot.send_chat_action(chat_id=update.message.chat_id, action=telegram.ChatAction.TYPING)
 		time.sleep(1)
